step2: refactoring + vite.config.js 추가
