+++ src/stores/label.js
@@ -1,6 +1,6 @@
-import {fetchData, isHexColor, removeItem} from "../utils.js";
-import AppState from "../libs/state.js";
-import {LabelModel} from "../components/label.js";
+import {fetchData, isHexColor, removeItem} from "@/utils.js";
+import AppState from "@/libs/state.js";
+import LabelModel from "@/components/label/model.js";
 
 /* Label Store를 object literal로 정의합니다. */
 const LabelStore = {
