+++ vite.config.js
@@ -1,0 +1,5 @@
+export default {
+  resolve: {
+    alias: [{find: '@', replacement: '/src'}],
+  },
+}
